build: add project dependencies
This is the initial dependencies of the project. It uses flask for the
web framework, along with packages that make certain tasks easier, more
efficient and more secure.
The main use-case for the project's core packages:
Flask: the project's web framework
MongoEngine: the ORM-like layer for interacting with MongoDB
Flask-Login: User session management for handling login and logout
functionality.
Flask-Bcrypt: For hashing passwords securely.
Flask-Mail: For sending email notifications (useful for push
notifications).
Flask-Talisman: For securing the app with HTTPS and security headers.
Flask-WTF: For handling forms in Flask with validation.
Flask-Cors: To handle Cross-Origin Resource Sharing, if the front-end
will be hosted separately.