add: configure application
app.py is the entry point of the application and makes use of `create_app`
to initialise the app. The `create_app` factory function which initialises
and configures the entire application based on the specified
environment. The environments are for "development", "testing" and
"production".
The environments' configuration differ according to the
attribute values stored in their respective class. A private .env file
stores the secret configuration values that should remain unknown. The
Mongo database connection settings are stored here as well.