update: provides defaults for flask-login
Flask-Login requires these four properties and the `UserMixin` class provides a generic implementation for defining these properties.
"is_authenticated()": that is True if the user has valid credentials or False otherwise. (I read on the teamtreehouse community that this attribute is sometimes alternated between property and method depending on the python environment used)
"is_active": that is True if the user's account is active or False otherwise.
"is_anonymous": that is False for regular users, and True only for a special, anonymous user.
"get_id()": a method that returns a unique identifier for the user as a string.