update: initialise flask login-manager
The base config class defines the new `REMEMBER_COOKIE_DURATION`
attribute to inform the flask app how long a stored cookie should live.
A LoginManager object is defined to implement login and logout opeations
by creating and storing `session` and `remember_me` keys to authorize a
user and access to protected routes. The app's `user_loader` method,
applied as a decorator, returns the user found in the database with each
request to a protected route. The user is stored in another variable,
called `current_user`, that can be accessed as an import from flask-login.