add: implement login view
A user's credentials for login are accessed via a request's json formatted data. If invalid data or incorrect user credentials are passed then appropriate error responses are sent as json data.
The login_manager object initialised in the app's factory function, currently in app/__init__.py, is set up to redirect a user from a protected route to the login route if the user is not yet logged in. Currently no forms, html pages or frontend is setup, so the user just receives a json redirect response.
The logout function is still not working as expected. A user's session details get removed according the flask-login properties like current_user, is_authenticated and get_id(), but when the already "logged out" user accesses a protected route, they get the same session details that should have been removed