Add route to update a user's details
Protect the route with @login_required. Access the logged-in user's
`User` object with the `current_user` object provided by flask-login
and expect json key/value pairs to indicate which user fields are ought
change. Allow only the user's 'username' field to be updated and ensure
the that the username is unique.

Currently, the 'username' field is the only non-critical field that
makes sense for a user to change aside from their
'personal_shopping_list' field (This field will be altered with other
routes). This is the reason that the user can only change their
username as of this commit.