Add route to change a user's password
Protect the route with `@login_required`. The logged-in user's details
are accessed with the `current_user` object provided by flask-login.
The route requires three keys to work as intended: 'current_password',
'new_password' and 'confirm_password'.

'current_password' needs to correctly the user's current password for
added authenticity aside from being logged in.
'new_password' is the value of the new password and needs to match
'confirm_password' to ensure that the user enters the correct password
they intend.