Add `admins` field to a household
Add `admins` field to the Household model and make it store references
to the `user` collection.

Doing this will allow the application to keep track of which users have
admin privileges in the household. In the future, these privileges may
be used to allow admin users to kick out other users from their
household.

Minor change:
Change the `users` filed to `members`
The `users` field is changed to `members` for a more appealing relation
to the term "household".