Add household blueprint and auth middleware
Initialize a Blueprint as `household_bl`.
Define middleware to implement authorization before accessing household
routes. In the middleware, ensure the logged-in user belongs to a
household. If they do, then validate that the household exists and has
the current user as one of its household members.

This middleware should be used for routes where only members of a
certain household is allowed access to the household. Therefore,
this middleware should be used for most of the household routes.

With this implementation, it doesn't require any sessions or ways to
keep the user as part of the household. Just like joining a group
on a messaging app, a user just needs to join a household once and they
won't be required to join again unless they are kicked out in some way,
eg. by a household admin.