Add route to create a household
Retrieve the 'name' and 'password' as json input from the
request. Validate them. Ensure that the 'name' is unique compared
to other existing household names. Hash the password and create a new
household object with the hashed_password, name and record the user
creating the household as its first admin and normal member.
Update the user to indicate the household they belong to.

Apart from the using the input to create a new household object and
Mongo DB document, the logged-in user and the new household form a
one-to-many relationship where a household can contain many users. The
household is however, only created when a user requires it and
therefore a household is dependant on users for it to exist in this
application. (A household does not exist if it doesn't have any
members)