Add route to display a household's details
Get the household object of the logged-in user. Return the household's
'name', 'admins', 'members' and 'created_at' field values.

The household's shopping_list field is not included as separate routes
will be dedicated to them. The 'password' field is confidential and the
'update_at' field is excluded as there is no use for that info as of
yet.

`current_user.household_id` seems to provide a household instance
instead of the actual household id. Therefore, no lookup in the
database is needed. A household object's `admins` and `members` fields
are a list of `user` objects. Since a `user` object value cannot be
parsed as json, new lists are created and filled with only a user's
username. These lists stand in place of these two list fields so that
they can be sent in the json formatted response body.