Add middleware only allowing household admin access
Check that the current_user object contains a value. (meaning - the
request comes from a logged-in user) Get the household that the user
belongs to and check if the logged-in user is an admin of the
household.

This middleware determines whether the request comes from a logged-in
user that is recorded as an admin in their household. This middleware
should be applied to household routes that only household admin members
can access, eg. routes that allow a user to add or remove members of a
household.
Applying this middleware reduces the need to continuously validate
whether the user trying to modify the household is an admin or not.