Add method to check the password of a household
Import werkzeug.security's `check_password_hash` function and use it in
a newly defined Household class method to match a household's hashed
password with a password passed as an argument.

It determines whether a password, matches the protected hash password
of a household. werkzeug.security's `generate_password_hash` function
is used to generate hashed passwords for a household. Thus, using their
checking function which knows the internal workings of the generating
function to be able to check the password against the hashed one makes
sense.