Add a route for user to join a household
Retrieve json expected input from the user with values for the `id`
and `password` of a specific household. Ensure that the specified
household exists and that its password matches the inputted password.
User is then added as a member of the specified household and its
household id is updated with the household that was joined.