Add a route to remove a household member
With the middelware used, only a household admin can access this route
to remove a user from the household. Checks are done to ensure that the
user intended for removal is part of the household and is not an admin.

With removal, the user's affiliation to the household is also removed.