Add route to promote a household member to admin
Get the user from the database based on the user_id provided in the
route. Ensures the user belongs to the household of the user making the
request and that the user is not already an admin. The user is then
added to the household's admin members.

As usual, an admin member should not modify other admin members.