Add route to add an item to house's shopping list
Create new file that handles all the views for a household's shopping
list, excluded from the household views itself.

Adding the new route:
Expect the item's name as json input and validate it. Create an instance
of the `ShoppingListItem`  model from `models.household` This model
defines the properties that an item for a household shopping list is
expected to have.

This route is needed so that item's with the desired format can be
added to a household's shopping list.