Add email_services file for email functionality
This file handles all functionality and interaction related to email
services that the application offers.

Define the `get_serializer` function that returns the app's token
serializer which is shared throughout the application.
Define the `generate_url_token` function which generates a secure token
using the app's secret salt value, stored in its configuration. Use the
app's shared serializer for generating the token.
Define the `send_confirmation_email` function that send an email to the
specified user's email address. The sent email contains a link for the
user to use to verify their email address.

The token generated by the `generate_url_token` function contains a
user's email information which will be used to validate which user
confirmed their email address for their account. A secret salt value
is used when generating the token and ensures that only those that use
this salt value with this shared serializer object can decode the
confidential data from the token. The app's shared serializer is used
before generating the token since this serializer in instantiated with
an extra layer of security.