Add a route to send an email to verify the email
Define the GET route and validate if the logged-in user requesting
this route already has their email marked as verified. If it does not,
then an email is sent to the email address of the user making the
request. Confirm whether or not the email was sent.

The route should not send an email to a user whose email is already
verified because it would be unnecessary and add clutter to the user's
email inbox.