Add a route for confirming a user's email address
Define a GET route that uses the token from the route and decodes it to
get the email of the user who is trying to confirm their address. Find
the user, using the email decoded from the token, and update that the
user's email adddress is valid.

The token is decoded using the app's shared serializer which uses an
extra layer of security. The serializer uses the same secret salt value, that
was used to generate the token, to decode the token. The token only
lasts for the specified period of time which is mentioned in the app's
configuration and this period of time differs according to the
environment the app is running.