Initialize mail object and shared token serializer
Register the blueprint meant for handling views that perform email
functionality for the app. Initialize a mail object with the app's main
flask instance. Add a serializer to the app's configuration, that s to
be shared across the app, in order to tokenize confidential data
securely.

The app's token serializer is shared instead of instantiating new
serializers when needed because the app's serializer uses an extra layer
of security by instantiating it with the app's secret key value. A
serializer generates a token secretely and only it has the essential
info that is needed to decode the token. Thus, you should use the same
serializer when encoding or decoding a token.

The extensions file is created to instantiate the extensions of the
flask app. The mail object could not be instantiated in __init__.py in
the app's factory function because trying to get this mail object from a
__init__.py from a view function would cause a circular import. Thus,
this extensions file is created to instantiate additional extensions to
flask, especially those that need to be used outside of __init__.py
(outside of the app's factory function meant for configuring the app)