Add configuration fields for mail and url tokens
Define mail and token fields in the base config class to be used for the
development and testing environments. Define mail and token fields for
production in the production config class.

Mailhog is used to capture all the emails sent to users in the
development and test environment. Their mail fields do not require much
authentic info because mailhog does not need them.