Fix users without a household not giving profile
Validate that the logged-user (`current_user`) has a value for their
household_id.

This validation needs to be done, otherwise if the user does not belong
to a house and the application directly tries to access the id property
of a NoneType (household_id), then an exception will be occur.